Refactor Browser components
